+++ docs/.vitepress/components/utils/IconBase.ts
@@ -1,5 +1,5 @@
 /* 修改自 vuepress-plugin-md-enhance
- * @src https://github.com/vuepress-theme-hope/vuepress-theme-hope/blob/main/packages/md-enhance/src/client/styles/tasklist.scss
+ * @src https://github.com/vuepress-theme-hope/vuepress-theme-hope/blob/main/packages/shared/src/client/components/IconBase.ts
  * @last-updated 2022-8-13
  */
 
@@ -30,71 +30,32 @@
  *
  */
 
- input[type='checkbox'].task-list-item-checkbox {
-  position: relative;
-  width: 0;
-  margin-right: 1.5em;
-  cursor: pointer;
+import { h } from 'vue'
 
-  html:not(.no-touch) & {
-    margin-right: 1.4em;
-    vertical-align: middle;
-  }
+import type { FunctionalComponent } from 'vue'
 
-  &::after {
-    content: ' ';
-
-    position: absolute;
-    top: 0;
-
-    display: inline-block;
-
-    box-sizing: border-box;
-
-    width: 1em;
-    height: 1em;
-    padding-left: 0;
-    border: 1px solid #ddd;
-    border-radius: 2px;
-
-    background-color: #fff;
-
-    text-align: center;
-
-    visibility: visible;
-
-    html.dark & {
-      border-color: #666;
-      background-color: #333;
-    }
-  }
-
-  &:checked {
-    &::after {
-      content: '';
-      border-color: var(--vp-c-brand);
-      background-color: var(--vp-c-brand);
-
-      html.dark & {
-        border-color: var(--vp-c-brand);
-        background-color: var(--vp-c-brand);
-      }
-    }
-
-    &::before {
-      content: '';
-
-      position: absolute;
-      top: 0.1em;
-      left: 0.35em;
-      z-index: 1;
-
-      width: 0.3em;
-      height: 0.7em;
-      border: solid var(--vp-c-white-soft);
-      border-width: 0 0.15em 0.15em 0;
-
-      transform: rotate(45deg);
-    }
-  }
+export interface IconBaseProps {
+  name?: string
+  color?: string
 }
+
+/**
+ * Icon Base Component
+ */
+export const IconBase: FunctionalComponent<IconBaseProps> = (
+  { name = '', color = 'currentColor' },
+  { slots }
+) =>
+  h(
+    'svg',
+    {
+      'xmlns': 'http://www.w3.org/2000/svg',
+      'class': ['icon', `${name}-icon`],
+      'viewBox': '0 0 1024 1024',
+      'fill': color,
+      'aria-label': `${name} icon`
+    },
+    slots['default']?.()
+  )
+
+IconBase.displayName = 'IconBase'
