+++ docs/.vitepress/styles/code-group.scss
@@ -1,73 +1,31 @@
-// /*  修改自 VuePress 2 (MIT)
-//   @src https://github.com/vuepress/vuepress-next/blob/main/packages/%40vuepress/theme-default/src/client/styles/code-group.scss
-//   @last-updated 2022-8-13
-//   @author Yuxi (Evan) You
-//  */
+.vp-code-group .tabs {
+  padding: 0;
+  background-color: inherit;
+  border-radius: inherit;
+  box-shadow: inherit;
 
-// /* code-group */
-// .code-group__nav {
-//   margin-top: 0.85rem;
-//   // 2 * margin + border-radius of <pre> tag
-//   margin-bottom: calc(-1.7rem - 6px);
-//   padding-bottom: calc(1.7rem - 6px);
-//   padding-left: 10px;
-//   padding-top: 10px;
-//   border-top-left-radius: 6px;
-//   border-top-right-radius: 6px;
-//   background-color: var(--vp-code-block-bg);
-// }
+  label {
+    font-size: 12px;
+    line-height: inherit;
+    padding: 4px 18px;
+    background-color: var(--vp-code-tab-bg);
+    border-bottom: solid 1px var(--vp-code-tab-divider);
+    /* border-left: solid 1px var(--vp-code-tab-divider); */
 
-// .code-group__ul {
-//   margin: auto 0 !important;
-//   padding-left: 0 !important;
-//   display: inline-flex;
-//   list-style: none !important;
-// }
+    &:first-of-type {
+      border-top-left-radius: 8px;
+      /* border-left: inherit; */
+    }
 
-// .code-group__li {
-//   margin: 0 !important;
-// }
+    &:last-of-type {
+      border-top-right-radius: 8px;
+    }
+  }
+}
 
-// .code-group__nav-tab {
-//   border: 0;
-//   padding: 5px;
-//   cursor: pointer;
-//   background-color: transparent;
-//   font-size: 0.85em;
-//   line-height: 1.4;
-//   color: rgba(255, 255, 255, 0.9);
-//   font-weight: 600;
-// }
-
-// .code-group__nav-tab:focus {
-//   outline: none;
-// }
-
-// .code-group__nav-tab:focus-visible {
-//   outline: 1px solid rgba(255, 255, 255, 0.9);
-// }
-
-// .code-group__nav-tab-active {
-//   border-bottom: var(--vp-c-brand) 1px solid;
-// }
-
-// @media (max-width: 640px) {
-//   .code-group__nav {
-//     margin-left: -1.5rem;
-//     margin-right: -1.5rem;
-//     border-radius: 0;
-//   }
-// }
-
-// /* code-group-item */
-// .code-group-item {
-//   display: none;
-// }
-
-// .code-group-item__active {
-//   display: block;
-// }
-
-// .code-group-item > pre {
-//   background-color: orange;
-// }
+@media (min-width: 640px) {
+  .vp-code-group div[class*='language-'],
+  .vp-block {
+    border-top-right-radius: 8px !important;
+  }
+}
