--- docs/.vitepress/styles/custom-container.scss
@@ -1,102 +1,0 @@
-.custom-block {
-  & .heimu {
-    &:hover a {
-      color: #fff !important;
-    }
-  }
-  html.dark & .heimu {
-    &:hover a {
-      color: currentColor !important;
-    }
-  }
-}
-
-.custom-block.tip .heimu {
-  background-color: var(--vp-custom-block-tip-heimu);
-  color: var(--vp-custom-block-tip-heimu);
-
-  a {
-    color: var(--vp-custom-block-tip-bg);
-  }
-
-  html.dark & {
-    background-color: var(--vp-custom-block-tip-heimu);
-    color: var(--vp-custom-block-tip-heimu);
-
-    a {
-      color: var(--vp-custom-block-tip-bg);
-    }
-  }
-}
-
-.custom-block.warning .heimu {
-  background-color: var(--vp-custom-block-warning-heimu);
-  color: var(--vp-custom-block-warning-heimu);
-
-  a {
-    color: var(--vp-custom-block-warning-bg);
-  }
-
-  html.dark & {
-    background-color: var(--vp-custom-block-warning-heimu);
-    color: var(--vp-custom-block-warning-heimu);
-
-    a {
-      color: var(--vp-custom-block-warning-bg);
-    }
-  }
-}
-
-.custom-block.danger .heimu {
-  background-color: var(--vp-custom-block-danger-heimu);
-  color: var(--vp-custom-block-danger-heimu);
-
-  a {
-    color: var(--vp-custom-block-danger-bg);
-  }
-
-  html.dark & {
-    background-color: var(--vp-custom-block-danger-heimu);
-    color: var(--vp-custom-block-danger-heimu);
-
-    a {
-      color: var(--vp-custom-block-danger-bg);
-    }
-  }
-}
-
-.custom-block.tsukkomi {
-  background-color: var(--vp-custom-block-tsukkomi-bg);
-  color: var(--vp-c-text-1);
-
-  a {
-    color: var(--vp-c-text-1);
-  }
-
-  html.dark & {
-    background-color: var(--vp-custom-block-tsukkomi-bg);
-    color: var(--vp-c-text-1);
-
-    a {
-      color: var(--vp-c-text-1);
-    }
-  }
-}
-
-.custom-block.tsukkomi .heimu {
-  background-color: var(--vp-custom-block-tsukkomi-heimu);
-  color: var(--vp-custom-block-tsukkomi-heimu);
-
-  a {
-    color: var(--vp-custom-block-tsukkomi-bg);
-  }
-
-  html.dark & {
-    background-color: var(--vp-custom-block-tsukkomi-heimu);
-    color: var(--vp-custom-block-tsukkomi-heimu);
-
-    a {
-      color: var(--vp-custom-block-tsukkomi-bg);
-    }
-  }
-}
