+++ docs/.vitepress/styles/index.scss
@@ -1,7 +1,7 @@
-@import './custom-container.scss';
-@import './code-group.scss';
-@import './mermaid.scss';
-@import './tasklist.scss';
+@use './container.scss';
+@use './code-group.scss';
+@use './mermaid.scss';
+@use './tasklist.scss';
 
 .vp-doc .custom-block blockquote > p {
   font-size: inherit;
@@ -20,7 +20,7 @@
   --vp-custom-block-danger-heimu: #f43f5d;
   --vp-custom-block-tsukkomi-heimu: #de22ff;
   --vp-custom-block-tsukkomi-bg: #de22ff24;
-  --vp-font-family-base: 'HarmonyOSHans', 'HarmonyOS Sans SC', -apple-system, BlinkMacSystemFont,
+  --vp-font-family-base: 'HarmonyOS Sans SC', -apple-system, BlinkMacSystemFont,
     'Segoe UI', Oxygen, Ubuntu, 'Fira Sans', 'Droid Sans', 'Helvetica Neue', Microsoft YaHei,
     sans-serif;
 }
